move syncState into engine
